MW-3 Improve Data Fetching
Fetching data in a call to `useEffect()` is an anti-pattern in React;
the problems with it, including the possiblity of race conditions and
stale closures leaving the sarch results out of sync with the search
term, are too numerous to list here.

For client-side React apps (ones bootstrapped with Vite, for example),
the canonical pattern is to use TanStack Query (formerly React Query) to
synchronize server state with client state.

This commit introduces a number of abstractions. They are not strictly
speaking necessary (or even desirable) in an app this small -- too many
abstractions introduced too early on makes a codebase harder to debug.
But for sufficiently large apps, colocating data fetching concerns and
api client concerns in separate modules is a useful pattern; I've
introduced it here because this is notionally a large app.