MW-4 Implement server-side filtering
Filtering the entire result set on the client has some benefits -- it
makes the UI more responsive when filtered because there isn't a network
request to the API -- *as long as the result set is small*. If there are
orders of magnitude more advocates than in the dummy data, not only will
the filtering performance degrade, but necessary improvements like
pagination will be harder to implement.

The particular filtering technique concatenates all an advocate's data
into a single string (using pipes `||` to prevent accidental matches
that span words) and tests that long string once against the search term
(all normalized to lowercase). This struck me as preferable to a lot of
different checks.